First attempt at fixing Repeats of nullable Matches.
I tracked down the source of the failures of the certain replacement rule
insertions without context (tests 36j, 36k, 36k-2). For one thing the problem
is tape priority dependent; it's also RTL/LTR dependent.

Writing additional tests and tracing them tracked the issue down to Repeats
involving nullable Matches. One of the simplest matches that fails is:
	(t2:e+M(t1>t2,ε|t1:h)){2}
with tape priority [t1, t2].
In this case, for example with a tape count of {t1:1, t2:3}, missing from the
results is: {t1:h, t2:eeh} for DIRECTION_LTR, {t1:h, t2:ehe} for DIRECTION_RTL.
This is because we fail to handle the situation where the h is not matched until
the second repetition. RepeatExpr.deriv() is failing to generate all of the
results.

We came up with a correction, but it is recursive. It is to take the deriv of the
concatenation of delta of the child and the repeat decremented by 1. The problem
with the recursive call of Repeat.deriv() is that it will recurse infinitely if
maxReps == Infinity. We added a guard against infinite recursion when
maxReps == Infinity, in which case it will fail; also, it can still recurse for
a long time before returning a result if maxReps is large.

New tests have been added to testMatchFrom. These are tests 30a to 30f-alt-2.
All of the -alt* tests fail without the above fix. Tests 30e-alt2 and 30f-alt-2
still fail because maxReps == Infinity. Test 30e-alt demonstrates the situation
when maxReps is large (set VERBOSE_DEBUG to trace what happens).