Implemented a non-recursive solution for the Repeats of nullable Matches issue from the previous commit.
Now we just return an "EMPTY_TOKEN" with the concatenation of delta of the
Repeat's child and the Repeat decremented by 1. generate now checks for the
"EMPTY_TOKEN" when producing output. CountExpr and CoutTapeExpr are also
appropriately modified.

The "EMPTY_TOKEN" handling is incomplete, so some tests involving Joins and
Filters are now failing. The "EMPTY_TOKEN" needs to be unique and handled
appropriately.

We are now getting all the correct results in the MatchFrom 30[a-f]-alt* tests,
but some results are now repeated.