Encapsulating execution environment as Env
(1) We were passing a bunch of environment objects (tapeNS, stack, opt) into each execution of deriv/delta, but most implementations don't actually make use of these, and they kind of act as a unit.  (The main exception to this being RenameExpr which affects the tapeNS, and EmbedExpr which affects stack.)

So now these have been wrapped in an edifice object Env.  Now the parameterization of all the query functions is much shorter, and the majority of implementations can just pass on the Env without having to think about what's inside it.

(2) Also fixed a bug in EmbedExpr.delta, where we forgot to wrap childNext in a new EmbedExpr.  This had effectively turned off recursion protection for multi-tape grammars.