Fixing up remaining multichar issues
(1) Had to decide yet again what to do with negation, but opted to just say "Negation is char-level still" (like last time) and push off the difficult decision until later again.

(2) Unit tests and multichar became an issue due to the way we set tape atomicity at construction, but different tests might cause a tape to be atomic or not (because they're filters in the end).

So atomic has to be set on the tape every time we collectAllVocab, not just the first time we call collectAllVocab.

(3) Testing for atomicity was causing the ol' slowdown in vocab collection in a big grammar, because it was over-visiting symbols.  Treating it more like collectVocab (caching by tapeName-symbolName keys) addresses this.