Small error-message improvements
Two small adjustments to syntax error handling:

(1) The sidebar interface wasn't suppressing duplicate error messages, now it does.

(2) Nested reserved words in headers (like "starts table") were erroneous but the errors were not reported.  Previously, the TstHeader was reporting this error, by seeing if its header was a ReservedErrorHeader, but this only works, of course, if that's the outermost header object. If the ReservedErrorHeader is inside of an otherwise valid header object, nothing is reported.  Now we collect errors from headers recursively and then report them.