Consolidating a new approach to error messaging
In the previous overhaul of error messaging, we consolidated around the idea that every object that can signal the programmer that there's an error (e.g., a TST or Grammar) retains a reference to the cell where that error message will be displayed.  So from sheet to TST to Grammar, we pass on Cell objects for this purpose.

However, we now want the possibility that errors are accumulated and displayed in cells higher up the "tree" (e.g., on a symbol saying "This symbol contains errors"), especially for when we've collapsed or hidden the actual content cells below it.

This violates the assumption we had previously made; there's no longer a unique cell that displays a message, and moreover cells themselves don't have any knowledge of the hierarchical structure necessary for this.  So we need to refactor towards an error mechanism that incorporates knowledge of the hierarchical structure of the program.

In the meantime, I've been liking the way "renameFix" works (and stays out of the way of everything else).  One transformation goes through the tree, detects a particular class of problems, AND transforms the tree to eliminate that problem. So I'm leaning towards error solutions more like that.

This commit does a variety of refactorings that will (eventually) coalesce into a more unitary error framework.

(1) So that we can do transform-style error passes to TSTs, there's now a TST transformation framework.  This one isn't a Visitor; I'm trying out a different style, where we just switch on class names.  (Why: During the last period of trying out lots of new Grammars, I got rather annoyed at having to write new transformation boilerplate every time, so I'm trying out a non-Visitor traversal.)  (However: I find myself forgetting to actually hook the transformation functions up to the main transformation function, something the type system would have warned me about if it were really a Visitor.  So maybe Visitor really is better here.)

(1a) Missing parameter and some inappropriate assignment errors are now TST transformations.

(2) All Grammar error messaging is now handled in an appropriate GrammarTransform.

(2a) Rename/Hide errors are now part of the RenameFix

(2b) There's now a MissingSymbol pass.  You might think this would be better inside the NameQualifier pass, and I don't disagree, but at the moment we can't do that because NameQualifier is run several times during project loading.  (We EXPECT missing symbols while the project is loading, that's our way of knowing what to import.)  But since error messagies are still handled as side effects, putting them as part of this pass fires them whether we mean it or not.  Eventually, when side effects are collected in another way, we can potentially merge these passes.

(2c) Unit tests are now handled as a transformation pass as well.  This eliminates the need for UnitTestGrammar objects to send their own messages, and also eliminates some of the special unit-test related methods in Grammar/UnitTestGrammar.

(2d) I haven't done anything with JoinReplace errors (e.g. having unexpected from/to tapes), just commented out those parts.  I'm not certain those errors were still valid (they seem to forbid epsilon to, which we can now handle), and we're not currently promoting JoinReplace as the right way to do rules right now anyway.  (It's not optimized the way JoinRule is now.)

(3) The only reason Grammars have Cells is error messaging, but we now need more than just Cell references.  However, I feel it's way too heavyweight for Grammars to have know about the whole tabular syntax system, or to keep a stack of relevant Cells that need to know the grammar's errors.

So I'm beginning to experiment with the idea that only special nodes in the Grammar tree know where they are, and we rely on their hierarchical relationships to inform the appropriate collection and display of error messages.  This is LocatorGrammar.  But I'm not using it for anything quite yet, that comes next.