De-genericizing GrammarTransforms
Not every GrammarTransform requires the exact same arguments (e.g., NameQualifier requires a stack of namespaces that no other transform does).  Previously I was handling this with a generic type GrammarTransform<T> where T was a tuple/record of every additional argument needed, and every GrammarTransform.transformX() and Grammar.accept() took an args:T argument.

This was a bit unwieldy, having to thread this type parameter through everything GrammarTransform touches (and it touches everything because it's a visitor).

But fortunately, it's rare that the values of this T actually change that much; for the most part, transformations just pass in the same args (and the same namespace argument) they were given.  Also, none of the Grammar.accept() methods ever actually use this argument, it's only the GrammarTransform.transformX() methods that use it.

Therefore, these args (as well as namespace) could just be members of the Transform object (or, looking at it another way, as captured variables in a Transform closure), and we create a new one in the limited circumstances when the args or namespace change.

This greatly simplifies the parameterization of everything the transforms touch.  But SURPRISE it's going to get worse again real soon.