Threading error accumulation through transforms
Transforms now return a list of error messages along with their results.  The housekeeping for it is a bit annoying, but ultimately, we want error messaging not to be a side-effect anymore, and if it's not a side-effect it needs to be part of the return value.

We're not actually using this return for anything yet, that comes next.