Fixing nested rule rename bug
This commit continues the refactoring of the error system, but in doing so found a bug that had been quietly fixed by RenameFix.

First the refactorings:

(1) Make LocatorGrammars report any errors that accumulate under them.  (This is probably temporary, since this is still reporting errors by side-effect.)  More .toGrammar() methods create them, too (we weren't creating quite enough of them before).

(2) RenameFix no longer triggers its own errors, they're now accumulated in the error returns and reported by LocatorGrammars.

(3) FilterTransform is now aware of LocatorGrammars and pattern matches appropriately.

Now the bug that was found and fixed:

Under the old error system, where error messages were carried through Cells, it was possible to have silent errors if the relevant Grammar had a DummyCell as its Cell.  I used this (well, abused this) to let the compiler do a few things that would have been invalid if a programmer had done them.

Silent errors like this are no longer possible in the new error system.  In making these explicit, I found a bug in ruleReplace2 where I was failing to hide/rename the .input tape, which could lead to an invalid tape structure when such rules were nested in a particular way.  RenameFix had been fixing this problem quietly and shooting the message off to the void (i.e. to DummyCell).

I had actually seen evidence of this problem: "Why is there sometimes an .ERR tape in non-erroneous grammars?  Welllll, regardless, everything seems to be working so I'll put off investigating."  That .ERR tape was RenameFix doing the rename that I had forgotten.

Anyway, now we do that rename explicitly in the code rather than ask RenameFix to do it, and meanwhile move RenameFix to earlier in the Transform chain, so that it's only detecting and fixing the programmer's renaming errors, not ones that we inadvertantly introduce.