Converting unit tests to new error system
The remaining grammar transforms now express their errors through the new error system, including the unit test transform.

Because the transform as stated won't have direct access to the addresses of unique fields, it can't display the related errors in the most-appropriate field anymore.  It now puts them in the leftmost field (same place it puts other errors).

We could still do this properly (with, say, a UniqueTestGrammar wrapped around the LiteralGrammar, and adding the error when the transform visits that), but no one is actually using uniques right now anyway, so no one will notice they're showing up in a different place ;)