Removing cell information from Grammars
The introduction of LocatorGrammars means that we can now associate elements in the Grammar tree with locations in the Sheets without necessarily having every Grammar object know where it is.

This commit removes the knowledge of cells from Grammar objects in general, except LocatorGrammars themselves and (for now) ReplaceGrammars (which still use their positions to name themselves).

Like the situation in the last commit with "unique"-tagged tests, there are some cases where our detection of an error is only done outside of the scope of the relevant Grammar (and thus outside the scope of its LocatorGrammar).  I put in a workaround accessor function location() that scans children for locations, it happens to work for the cases where we use it, but it probably won't work in the larger picture and we should think of better solutions in the future.

This commit also fixes an efficiency bug, where calling a TstTransform twice in the same call was going exponential and leading to impossibly long compile times.