Removed Cell reference from ReplaceGrammar
Aside from LocatorGrammar (which obviously requires a Cell reference), the only Grammar remaining that had a Cell was ReplaceGrammar, which used it to name itself in the absense of a client-provided name.

But this Cell reference wasn't really necessary; the Cell reference in HiddenGrammar and ReplaceGrammar was just to solve a bug that has been long since solved another way.  HiddenGrammar had already switched to incrementing an index on initial construction and passing it along to its clones.  So ReplaceGrammar does that too.