Errors during transformations now effectless
Part of this big error refactoring is to move away from doing errors as side-effects, for several reasons.

This commit does away with errors-as-side-effects within GrammarTransforms.  Rather than LocatorGrammar firing off the error message and then returning no errors, it just localizes its children's errors and returns the localized errors.

The caller of the transform can then decide whether to fire off those messages or not.