Simplifying symbol resolution passes
When transformation errors were side effects, NameQualifier and missing-symbol error handling had to be separate -- NameQualifier couldn't do error messaging itself because we run it many times.

Instead, UnresolvedEmbedGrammars served as a bridge: when any UnresolvedEmbedGrammars survived the NameQualifier step, we used them to either:

(a) send an error to the programmer, and then transform them into innocuous EpsilonGrammars so the grammar is valid, or
(b) collect unresolved names for sheet loading

Now NameQualifier can do all of that itself, because errors are just return values; the caller can just toss them if they're not appropriate to send to the programmer.

Because of this change, UnresolvedEmbedGrammars no longer did much of anything, and their lifetime was extremely short.  (They only existed in between TST.toGrammar and NameQualifier.tranform, but those calls happen immediately next to each other.)  So I got rid of them; once again only EmbedGrammars exist.

(NB: This introduces one complication with EmbedGrammars that we have to be careful of.  Like the former UnresolvedEmbedGrammars, EmbedGrammars don't yet know their namespaces at the time of their construction, now they only know it after NameQualifier resolves them.  But because of this, we have to be careful not to call anything that needs EmbedGrammar.namespace in between TST.toGrammar() and this resolution, like .calculateTapes().  Either that or we have to let EmbedGrammar silently fail and have no tapes when it doesn't have a namespace.  But I'd like to avoid silent failures, even ones we expect, so let's just be careful instead.)

Also, messages are classes now, so that we can subclass them and thus send along arbitrary, error-specific information.  In this case, MissingSymbolError has a member .symbol, so that sheet loading can just grab unresolved symbol names straight from the error message.