Improvements to assign & op handling
Several improvements, changes, and bugfixes to the handling of assignments and operators.

(1)  It used to be that every TST used the child/sibling structure to handle its children.  (That is, parent nodes only have one child, and when a child is added, that becomes the only child and the former child becomes its sibling.)  That makes for a nice binary structure for ops like join, replace, etc. but was a bit odd for namespaces and assignments.  Sequences of assignments don't conceptually have that kind of binary structure, you think they'd just be a list and their parent namespace would know about them.  But nope, they were binary just like join and replace.

Handling assignments this way meant that the namespace doesn't directly know about them, and we had to have a recursive algorithm assignToNamespace() to assign things to namespaces properly.  This algorithm was a bit too clever, and when it came time to separate out thing like error-handling passes into transformations, it would have been unwieldy to duplicate how it works, just to traverse the children of namespaces.

So namespaces are just flat again, they know their children directly.

(2) Unlike other binary ops, unit tests had an exceptional syntactic possibility: they can be placed directly underneath an assignment (rather than under a "table:") and still work, like so:

MAIN:, embed,     text
     , verbstem,  a

test:, text,      subj, tense, root
     , ninapenda, 1,    pres,  pend

Not that there was any special code to handle this; it just works according to the existing syntax/semantics.  Unit tests, after all, aren't accessed through names, they execute through a special tree traversal.  So the fact that they're not assigned to MAIN here doesn't matter in the least.  And during normal execution, tests are semantically transparent.  So this program is semantically identical to the program where the test is indented and sibling to a "table:" op.

However, this identity wasn't true in general for binary ops.  A programmer might think that because this works for "test", it'll also work with superficially-similar operators like "replace", but it doesn't.  That replace would never get assigned to anything (unless it happened to be the last thing in a page).

I decided to generalize this identity for consistency: now, any binary operator like test, replace, join, etc. can be placed in this position, and will be treated as scoping under MAIN, just like if it were indented placed under a "table:" within the scope of MAIN.  AdjustAssignmentScope is a new TST transformation that handles adjusting the scopes as necessary.

(3a) I had failed to properly hook up TstReplaceTape (the TST representing new-style replacements to MissingParamsTransformation), and we weren't testing for it.  (All the tests were for old-style replacements.)  This commit addresses both.

(3b) While doing the above, I found some other old-style replace tests that had no new-style counterparts, and made tests for those too.