Threading errors through TstTransforms
(1) TstTransforms are called multiple times (because they're a prereq to name qualification, and name qualification happens multiple times), so if errors are side-effects, the errors get multiply reported.  So this commit does the same thing to TstTransforms that was previously done to GrammarTransforms: makes the errors part of the return value instead.

(2) Since they were getting long, moved them each to their own file.

(3) The sequence of TstTransforms lived in two different places -- in sheets.ts during sheet loading, and in Interpreter for all other purposes.  So when this order changed, it had to be changed in two places.  So now there's a composite transfomer, TstTransformAll, that encapsulates that order.