Error handling as monads
Now that we don't want to do error handling as side-effects (because that was a mess), they have to be return values.  Every function that can trigger an error needs to have an additional return value for any errors, and same with every caller of it.  That means every caller has to engage with the collection and concatenation of error lists as well.

But this ALSO became a bit of a mess; it's way too easy to just drop errors somewhere.  And this isn't, fundamentally, what we want to be thinking about algorithmically; these functions should spend most of their time describing how the input->output mapping works, not doing the donkey work of concatenating error lists.  We want to do as much as possible without handling the error return values explicitly.

To do that requires going up one abstraction level beyond returning a tuple.  Rather than returning a [T, Msg[]] tuple, we wrap return values in some sort of container, call it Result<T>, and let that handle the boilerplate about error collection.  We also want to avoid destructuring it (because then we're back in the situation where we have to think about threading error handling through an algorithm again); rather, we should have some sort of map function that lets us apply functions to its item and gives us a new Result, with any additional errors caused by the application of that function handled in the background.

This, whether we like it or not, is a Monad.  I know our coding standards say not to go too functional to the point where it becomes obscure, and nothing is more obscure than monads... But our error handling was ALREADY monadic, those [T, Msg[]] returns were monads, we just weren't automating away the boilerplate.  (There's a quote I can't manage to dig up right now, like "Category theory is to programming what chemistry is to baking; it's there regardless of your awareness of it.")

Anyway, Result<T> is an object that:

  * combines item:T and msgs:Msg[]
  * has a bind() method that takes a function, applies it to the T, and returns the resulting value wrapped in a new Result, with any resulting errors flattened into the .msgs.
  * has a .msg() method that adds additional messages to the msgs.  (TSTs and Grammars also now have a helper method .msg() as well, that lifts these into Results.  If you just want a result, the easiest way is just to call x.msg() with no arguments.)
  * For the most part, we should try to use Results just by using .bind() and .msg().  But there's also a method .destructure() that will give you back the [T, Msgs[]] tuple that we were using before.

Btw, originally the .item and .msgs members of Result were public, and so you didn't have to explicitly destructure.  But this was actually trouble.  If you can get at x.item any time you want, you can do stuff with it, use it to make some new Result and return that, and lose track of x.msgs.  Forgetting about messages is the very thing we want to avoid.  When we explicitly destructure, like

   const [item, msgs] = result.destructure()

and then forget about the msgs, msgs is at least dehighlighted in the IDE as an unused value.  (This isn't perfect, because this dehighlighting goes away so long as we use it ONCE, but even so this has been really handy at reminding me to handle the msgs.)  So .item and .msgs are now protected; destructuring is annoying but it's annoying with a purpose, to force us to think about the fact that we're doing it and remind us about restructuring later.