Fixes to deriv for RepeatExpr, JoinExpr, FilterExpr, and IntersectExpr.
The fix to RepeatExpr.deriv() in the previous commit (e83a0412) limited the
yielding of the EpsilonToken too much. In fact, it does need to be yielded even
when a specific target character is specified because that character could be
yielded by a subsequent iteration of the repeat.

I expanded the tests for this quite substantially, putting them all in testRepeat
since they are all related to repetitions of nullable matches.
The tests are based on the expression:
	(t2:e+M(t1>t2,ε|t1:h)){2}
which forCountTape(3), generates the following results:
    {t2: 'ee'},
    {t1: 'h', t2: 'ehe'},
    {t1: 'h', t2: 'eeh'},
For each test family, we test generation (as shown above), joins in both
directions with each of the three results (variants of the above listed ones),
filters with each of the three results, and intersections in both directions
with each of the three results.
The test families added are:
	(t2:e+M(t1>t2,ε|t1:h)){2}
	(t2:e+M(t1>t2,ε|t1:h)){2} + t1:x
	(t2:e+M(t1>t2,ε|t1:h)){2} + t2:x
	(t2:e+M(t1>t2,ε|t1:h)){2} + (t2:e+M(t1>t2,ε|t1:h)){2}
	((t2:e+M(t1>t2,ε|t1:h)){2}){2}
	((t2:e+M(t1>t2,ε|t1:h)){2})*
	((t2:e+M(t1>t2,ε|t1:h))*){2}
The other side of many of the joins, filters, intersections use repeats too.
In all 127 tests of the above sort were added, plus a few others.

All tests now pass except for some generating duplicates resulting in the number
of results test failing, until we think up another combination that fails.
Also, when I say everything passes, that means with the settings:
	export const BITSETS_ENABLED: boolean = false;
	export const DIRECTION_LTR: boolean = false;
The implementation for bitsets is not yet complete, and LTR needs to be tested.