Step towards unifying transformation frameworks
Our transformation frameworks (Tst- and Grammar-) work in slightly different ways (which is okay for now), but they also have slightly different and incompatible APIs.  It would be nice if eventually everything were transforms and we could just say "Here's the order they come in, execute it."

Changes:

(1) Unifying them in that way is complicated by the fact that these transformations operate on different types, and potentially change the type of the transformed object.  (E.g., toGrammar would be a transformation that changes its object's type.)  So it's difficult to do what we'd been doing in a type-safe way, just storing them in an array and looping over that because they don't have any common interface, and the resulting object would change type sometimes.

So now they're both subclasses of a genericized class Transform<T1,T2>, which has a type-safe composition operation.  This lets us treat sequences of transforms as if they were single transforms, so long as the output/input types are compatible.

(2) I threaded an env parameter through both Tst- and GrammarTransforms.  I'd been meaning to do this for a while, there are some decisions that Transforms needed to make that depend on things like whether we're LTR or RTL, and without something like the env that's a global.

I don't love threading params through transforms in this way because it affects a ton of methods and the env rarely actually changes.  Just a few commits ago I took a several params out of GrammarTransform and made them members instead.  But I feel like "env" is different; putting it as a param makes it more similar to how delta/deriv work.

(3) Rather than redo the way TstTransforms work internally to be more like GrammarTransforms or vice-versa, there's now just a wrapper class that wraps around a GrammarTransform but has the API of a TstTransform.

Eventually it'd be nice to make them work more similarly inside as well.  I have the hunch that a Visitor pattern would become very awkward here, generalized to the entire framework.  Transforms with different types would all have to have different XTransform interface and IdentityTransform ancestors, and any new class would have to be added to every transformation A=>B where the class extends/implements A.  This isn't too bad NOW, but it could become a mess as the chain gets more complicated.