Better named param handling
The way were were handling named params was pretty convoluted, now it's slightly less so.

(1) We had a whole grammar-generation subsystem within params to handle the fact that rows might contain differently-named params (which therefore shouldn't concatenate etc.).  We can just do that within TstRow; all the information we need to segregate cells by their param name is available as we're adding them to the row in the first place.

(2) Also, we had never actually flagged unwanted named params as errors in the header cell where they occur.  We only ever did that when constructing the grammar, meaning each cell underneath was an error.  That's not usually how other errors work; when the header is erroneous the header is marked as an error and the cells beneath it are only warnings.  So now the header itself is an error.

(The content beneath it isn't a warning, at the moment.  I'll get to that later.  This is just a symptom of a bigger problem, that even if we change out the header to a TstEmpty or a new kind of error header during a transform, it's already been paired with the content, so from the content's point of view it's the same, nothing we do to the headers actually changes the grammar itself.  So really, we need to pair up header and content only AFTER header errors are handled.)

(3) This change doesn't yet address missing required headers, like to/from in replace.  (I guess this isn't entirely semantically required -- "to" and "from" can be epsilon -- but it'd still be nice.  I think it's more likely that leaving out the "to" column is a mistake on the programmer's part, rather than the programmer deliberately using that to signal that they want "to" to be epsilon.)

(4) Also changed: you used to be able to nest "table:" under things like "replace:"; now this messes up the param checking because "replace" expects to/from/pre/post and "table" expects none of these.