Rename ParamRow -> TstRow
Oops, this was meant to be part of the previous commit.  (I had originally implemented the new TstRow as a different class, and only later realized that it could replace TstRow everywhere.  So I got rid of the old TstRow and renamed ParamRow back to TstRow, but forgot to git add that.)