Simplifying header->grammar interaction
One of the biggest complications of turning sheets into Grammars is that the nature of the horizontal cell relationship (i.e. a cell coming after another in a row) depends on that cell's header.  The default interpretation is that they're in sequence with each other, but "rename", "hide", "equals", "starts", "ends", and "contains" create a non-default relationship with the previous cell(s).

The way we used to do this ended up pretty convoluted: rows were handled as rightward linked lists of TstHeadedCells (i.e. they knew their previous cell), which then passed that .prev argument into .header.toGrammar(), which had a special parameterization to handle this.  This made it so the TST tree didn't actually have to know which headers did what, but it eventually got pretty difficult to wrap one's head around (due in part to there also being complications with regex cells).  But at the time, we had nowhere ELSE to do this complicated bit, .toGrammar() was the only pass over TSTs/headers.

Now we have multiple passes over TSTs, which gives us options like rescoping.  Now:

(1) TstHeadedCells no longer keep a link to their previous, and no longer pass it to .header.toGrammar().

(2) Instead of keeping a rightward linked list of TstHeadedCells, TstRow keeps a convenience object called a TstSequence.  It's got the default semantics of being a SequenceGrammar of its children.

(3) During a special TstTransform pass, we take a look at those children, and if they have one of the special headers rename/hide/equals/etc., we pop the previous child and make a special TstComponent (e.g. TstRename) that DOES keep a reference to its prev.  (In other words, the only Tsts that keep a prev are those that genuinely scope over it in the semantics.  We're no longer doing simple sequencing by prev, that's now flat.)

(4) Rename and hide used to have different scope rules than equals/starts/ends/contains -- they scoped over EVERYTHING to their left, not just the previous cell.  However, I'm not sure anyone was actually using this functionality, and it's another thing to remember.  (Okay, technically, it was the same scope as concatenation, but I don't think any of our actual programmers thought of concatenation in that way so as to analogize that to rename/hide.)  Anyway, new syntax rule: now rename/hide have the same scope as equals/starts/ends/contains.

In the big picture, this refactoring is in line with the general trend of refactorings lately. We used to have custom recursive methods that moved over common structures (e.g. sib/child) that didn't necessarily reflect each node's syntax/semantics, and passed in whatever information was locally missing.  Now, we try to have a richer and less homogenous structure, traversals have fewer custom parameters, and primarily operate by replacing tree nodes with other tree nodes.