Tst.toGrammar() now also returns Results
toGrammar, toParams, and toParamsTable all use the monadic error handling now.  This was a bit tricky to do but there's not actually anything of note about the changes, it's all along similar lines as the previous commits.

There are now a few more convenience methods on Result, .log() to log all messages to the console, and .localize() to assign a position to unpositioned errors.  (So what LocatorGrammar did, but factored out so that things other than LocatorGrammar can do this.)

I think, now, that there are no actual side-effect errors anywhere after the construction of the tabular syntax tree.  All of the remaining side-effect errors are during the sheet-parsing process.