Refactoring op construction
This commit slightly delays the construction of ops, in order to have more information available during their construction.

Now, rather than the sheet parsing algorithm knowing about op parsing, all it knows is that ops end with ":", and it creates a TstAnonymousOp corresponding to that.  In a subsequent step, we go through the TST, parse the op in the cell, and turn it into whatever it is.