Separating out namespace & op construction
It'd be nice if, by the time things like TstBinaryOp and TstReplace and company are created, we can verify that they have all the structural and named arguments that they need in order to be well-formed.  That was part of the reason for the previous refactoring, to delay their creation to the point where more information about their arguments was available.

However, there's a bit of a catch-22 if we do it all in one pass: TstNamespace needs to restructure its children and their arguments, so it should be done before their transformation from TstAnonymousOp->WhateverTheyActuallyAre, so they can know about those arguments.  However, to restructure them, TstNamespace needs to know what they are, not just that they're TstAnonymousOp, which means it needs to occur after that transformation.

So this commit:
  (1) moves op parsing back into sheet parsing (so it's already been done)
  (2) gives TstAnonymousOp an op member (and renames it to TstOp because it's no longer so anonymous)
  (3) handles the creation of TstNamespace first, along with restructuring its children
  (4) delays the creation of TstBinaryOp, TstReplace, etc. until later.

Two more minor changes:
  (1) To distinguish it from TstTableOp, I renamed TstTable to TstGrid, with the intention of eventually making it semantically more neutral and moving the interpretation of it into TstTableOp.
  (2) Also, the responsibility for wrapping sheet namespaces in assignments before their addition to TstProject is now in sheets.ts.  This means there's nothing for TstProject to do anymore, distinct from TstNamespace, so it's gone, there are only TstNamespaces now.