Remove last of old error system
There were only a few places left that used the old error system; this commit removes them, and also the last of the support structure for it.  SheetComponent.message() and TstComponent.message() are gone, SheetCell is gone (all it did was forward messages), and Cell is now just a plain class instead of an interface.  DummyCell is also gone (there was only one place we still used it, as the "cell" of a TstProject).

Other small changes:

(1) .handleMsgs() was renamed .msgTo() and can take an array argument, to which the messages are pushes.  So `x.msgTo(msgs)` would be equivalent to earlier `x.handleMessages(m => msgs.push(m))`.

This way, if you're just collecting messages in some array somewhere (which is pretty common), you don't have to destructure a result (including naming the messages something).

(2) A convenience class for Result<void>, called (you guessed it) ResultVoid.  This may seem weird but lifting void to a result gives us access to useful message-handling function.  Consider the situation where we want to return messages from a void function like Tst.addChild(), then push them to msgs:

    msgs.push(...tst.addChild(x));
vs.
    tst.addChild(x).msgTo(msgs);

The former approach was resulting in illegible code; algorithmically important things were getting visually lost because they were nested inside message handling code.