Unifying op construction & param checking
(1) As mentioned, it'd be nice to make sure that ops had all their arguments before we actually instantiate them as TSTs.  That way, if an op is ill-formed, we can just not create it, rather than create it and then any time we do anything with it, check whether it's missing params or something.

Structural argument checking (e.g. sibling and child) and valid assignment checking are now in CreateOps, before the relevant ops are created, rather than putting them off until a later pass.

This eliminates MissingParamsTransform and removes a lot of the work from InvalidAssignmentTransform.  The latter is simplified considerably (by taking the POV of the parent and checking the children, rather than taking the POV of the child and having to pass in a parental figure).

(2) It moves the details of Op->Tst transformation into (or back into) CreateOps, rather than being members of the Op hierarchy.  They felt strange down in src/ops.ts, because it was a case of something in src/transforms using an algorithm specified largely in src/.

(3) I was getting confused by both TstTransform and TstComponent having methods both named .transform() that mutually recurse.  In TstTransform.transform(t) do I want to call this.transform(t, env) or t.transform(this, env).  Both compile, but one recurses forever...

So now TstComponent.transform is called .mapChildren(), which is basically what it does, it maps the TstTransform to each of the Tst's children then clones the Tst with those new children.

(4) "table:" can't be the immediate child of unit tests anymore, just like the recent change to (if I remember correctly) replacements.