Sheet.toTST now a transform
(1) The parsing of sheets and their conversion to a TST is now in a transform, and concatenated with the TST transforms rather than being run as a separate step.  (This is the first time we've actually done a heterogenous concatenation of Transforms and it works, woo.)

(2) addSheet has been moved to interpreter.ts.  Although it feels like something that should be in SheetProject, it was actually a bit weird there because it had to know roughly half the interpretation pipeline.  (It has to be able to shuttle a page all the way to name qualification, in order to identify possible source files that need loading.)  Interpreter is where all that stuff usually lives, so now it's in that file.